src/server/requirements.txt
azure-functions==1.21.3
fastapi==0.115.6
pydantic==2.10.4
pydantic-settings==2.7.1
azure-ai-projects==1.0.0b7
azure-identity==1.19.0
openai==1.58.0
